update the init function
